added nodes

+++ BunnyHop/BhDefine/Functions/event/Statement/onStatMovedToWS.js
@@ -4,7 +4,9 @@
 	let loopCtrlStatList = [];
 	bhThis.findSymbolInDescendants(1, true, loopCtrlStatList, 'BreakStat', 'ContinueStat');
 	for (let i = 0; i < loopCtrlStatList.length; ++i) {
-		bhNodeHandler.deleteNode(loopCtrlStatList[i], bhUserOpeCmd);
+		const loopStat = loopCtrlStatList[i].findSymbolInAncestors('LoopStat', 1, true);
+		if (loopStat === null)
+			bhNodeHandler.deleteNode(loopCtrlStatList[i], bhUserOpeCmd);
 	}
 	
 })();
